Extract hash and keyholder handlers
