+++ src/utils/hash.js
@@ -1,0 +1,13 @@
+export async function generateHash(text) {
+    if (!text) return null;
+    try {
+        const encoder = new TextEncoder();
+        const data = encoder.encode(text);
+        const hashBuffer = await crypto.subtle.digest('SHA-256', data);
+        const hashArray = Array.from(new Uint8Array(hashBuffer));
+        return hashArray.map(b => b.toString(16).padStart(2, '0')).join('');
+    } catch (error) {
+        console.error('Error generating hash:', error);
+        return null;
+    }
+}
